+++ docs/changelog.md
@@ -37,6 +37,7 @@
 ### ATT&K Workbench version 1.3.2
 
 #### New Features in 1.3.2
+-   Added admin interface for creating and managing teams
 -   Added ability to search objects by one or more users.
 -   Added a new user profile page where logged in users can view their recent activity across the Workbench and view/edit their display name.